requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
Requests==2.31.0
